+++ app/src/main/java/com/example/cloudnotify/data/model/local/HourlyWeather.kt
@@ -11,8 +11,8 @@ data class HourlyWeather(
 val dt: Int,  // Original timestamp (as received from API)
 val hour: String,  // Formatted hour (e.g., "15:00")
 val weatherDescription: String,  // Weather description
-val temperature: Double,  // Temperature for the hour
-    val tempMax: Double,
-    val tempMin: Double,
+val temperature: Int,  // Temperature for the hour
+    val tempMax: Int,
+    val tempMin: Int,
 val icon: String  // Weather icon for the hour
 )
